[modify] バーコードスキャンを調整
